# doc/OpenAPI 3.1.yaml
openapi: 3.1.0
info:
  title: Project Chimera - S12 Gateway API
  version: 0.3.0
servers:
  - url: https://chimera.example.com/api
security:
  - bearerAuth: []
paths:
...
  /reconcile:
    post:
      summary: Reconcile orders/positions with exchange
      operationId: reconcile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReconcileRequest" }
      responses:
        "200":
          description: Reconcile summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReconcileResponse" }
...
  /positions/manage:
    post:
      summary: Manage positions (trail SL / partial TP / add if ok)
      operationId: positionsManage
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PositionsManageRequest" }
      responses:
        "200":
          description: Manage result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionsManageResponse" }
  /labels/backfill:
    post:
      summary: Backfill labels (12/24/36h)
      operationId: labelsBackfill
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabelsBackfillRequest" }
      responses:
        "200":
          description: Backfilled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabelsBackfillResponse" }
  /autopsy/{trade_id}:
    post:
      summary: Generate autopsy report
      operationId: autopsyGenerate
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: trade_id
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutopsyRequest" }
      responses:
        "200":
          description: Report meta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutopsyResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ReconcileRequest:
      type: object
      properties:
        mode: { type: string, enum: [ALL, ORDERS, POSITIONS], default: ALL }
        dry_run: { type: boolean, default: false }
        orphan_policy: { type: string, enum: [RECLAIM_IF_SAFE, CONSERVATIVE], default: CONSERVATIVE }
        time_window_h: { type: integer, minimum: 1, maximum: 168, default: 72 }
      required: [mode]
    ReconcileResponse:
      type: object
      properties:
        summary:
          type: object
          properties:
            checked_orders: { type: integer }
            checked_positions: { type: integer }
            adopted: { type: integer }
            closed: { type: integer }
            fixed: { type: integer }
        items:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [ORPHAN_ORDER, ORPHAN_POS] }
              action: { type: string, enum: [ADOPTED, CLOSED, FIXED] }
              ref: { type: string }
...
    PositionsManageRequest:
      type: object
      properties:
        symbols:
          type: array
          items: { type: string }
          minItems: 1
        actions:
          type: array
          items: { type: string, enum: [TRAIL_SL, PARTIAL_TP, ADD_IF_OK] }
          minItems: 1
        dry_run: { type: boolean, default: false }
      required: [symbols, actions]
    PositionsManageResponse:
      type: object
      properties:
        managed: { type: array, items: { type: object } }
        errors: { type: array, items: { type: string } }
    LabelsBackfillRequest:
      type: object
      properties:
        horizon_h: { type: integer, enum: [12,24,36], default: 24 }
        from_ts: { type: integer }
        to_ts: { type: integer }
      required: [horizon_h]
    LabelsBackfillResponse:
      type: object
      properties:
        updated: { type: integer }
        errors: { type: array, items: { type: string } }
    AutopsyRequest:
      type: object
      properties:
        force_rebuild: { type: boolean, default: false }
    AutopsyResponse:
      type: object
      properties:
        trade_id: { type: string }
        report_id: { type: string }
        url: { type: string }
...